engine/requirements.txt
httpx>=0.27
pyyaml>=6.0
rich>=13.0
sentence-transformers>=3.0
torch>=2.0
numpy>=1.24
python-slugify>=8.0
